+++ packages/react-native-worklets/Common/cpp/worklets/Tools/WorkletsVersion.h
@@ -13,7 +13,6 @@ namespace worklets {
 
 std::string getWorkletsCppVersion();
 void injectWorkletsCppVersion(jsi::Runtime &);
-bool matchVersion(const std::string &, const std::string &);
 void checkJSVersion(
     jsi::Runtime &,
     const std::shared_ptr<worklets::JSLogger> &);
